+++ src/MvcCore/Ext/Controllers/DataGrids/Configs/IColumn.php
@@ -17,65 +17,65 @@ interface IColumn {
 	
 	/**
 	 * Get data grid model property name.
-	 * @return string|NULL
+	 * @return string|null
 	 */
 	public function GetPropName ();
 
 	/**
 	 * Set data grid model property name.
-	 * @param  string|NULL $propName
+	 * @param  string|null $propName
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetPropName ($propName);
 	
 	/**
 	 * Get grid heading human readable text.
-	 * @return string|array|NULL
+	 * @return string|array|null
 	 */
 	public function GetHeadingName ();
 	
 	/**
 	 * Set grid heading human readable text.
-	 * @param  string|array|NULL $headingName
+	 * @param  string|array|null $headingName
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetHeadingName ($headingName);
 	
 	/**
 	 * Get grid heading title attribute, displayed on heading mouse over.
-	 * @return string|array|NULL
+	 * @return string|array|null
 	 */
 	public function GetTitle ();
 	
 	/**
 	 * Set grid heading title attribute, displayed on heading mouse over.
-	 * @param  string|array|NULL $title
+	 * @param  string|array|null $title
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetTitle ($title);
 
 	/**
 	 * Get database column name.
-	 * @return string|NULL
+	 * @return string|null
 	 */
 	public function GetDbColumnName ();
 	
 	/**
 	 * Set database column name.
-	 * @param  string|NULL $dbColumnName
+	 * @param  string|null $dbColumnName
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetDbColumnName ($dbColumnName);
 
 	/**
 	 * Get URL column name when sorting or filtering.
-	 * @return string|array|NULL
+	 * @return string|array|null
 	 */
 	public function GetUrlName ();
 	
 	/**
 	 * Set URL column name when sorting or filtering.
-	 * @param  string|array|NULL $urlName
+	 * @param  string|array|null $urlName
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetUrlName ($urlName);
@@ -85,7 +85,7 @@ interface IColumn {
 	 * Boolean to enable/disable anchor url for filtering
 	 * (`TRUE` by default) or string to declare row instance
 	 * model public method to generate custom URL.
-	 * @return bool|string|NULL
+	 * @return bool|string|null
 	 */
 	public function GetUrlHelper ();
 
@@ -94,20 +94,20 @@ interface IColumn {
 	 * Boolean to enable/disable anchor url for filtering
 	 * (`TRUE` by default) or string to declare row instance
 	 * model public method to generate custom URL.
-	 * @param  bool|string|NULL $urlHelper
+	 * @param  bool|string|null $urlHelper
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetUrlHelper ($urlHelper);
 
 	/**
 	 * Get datagrid column index, starting with `0`, optional.
-	 * @return int|NULL
+	 * @return int|null
 	 */
 	public function GetColumnIndex ();
 
 	/**
 	 * Set datagrid column index, starting with `0`, optional.
-	 * @param  int|NULL $columnIndex
+	 * @param  int|null $columnIndex
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetColumnIndex ($columnIndex);
@@ -115,14 +115,14 @@ interface IColumn {
 	/**
 	 * Get default sorting definition with values `ASC | DESC` 
 	 * or `TRUE | FALSE` to enable/disable sorting.
-	 * @return string|bool|NULL
+	 * @return string|bool|null
 	 */
 	public function GetSort ();
 	
 	/**
 	 * Set default sorting definition with values `ASC | DESC` 
 	 * or `TRUE | FALSE` to enable/disable sorting.
-	 * @param  string|bool|NULL $sort
+	 * @param  string|bool|null $sort
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetSort ($sort);
@@ -142,7 +142,7 @@ interface IColumn {
 	
 	/**
 	 * Get property type(s), necessary for automatic formating.
-	 * @return \string[]|NULL
+	 * @return \string[]|null
 	 */
 	public function GetTypes ();
 	
@@ -167,39 +167,52 @@ interface IColumn {
 	
 	/**
 	 * Get property automatic parser arguments.
-	 * @return array|NULL
+	 * @return array|null
 	 */
 	public function GetParserArgs ();
 	
 	/**
 	 * Set property automatic parser arguments.
-	 * @param  array|NULL $parserArgs
+	 * @param  array|null $parserArgs
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetParserArgs ($parserArgs);
 	
 	/**
 	 * Get property automatic formating arguments.
-	 * @return array|NULL
+	 * @return array|null
 	 */
 	public function GetFormatArgs ();
 	
 	/**
 	 * Set property automatic formating arguments.
-	 * @param  array|NULL $formatArgs
+	 * @param  array|null $formatArgs
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetFormatArgs ($formatArgs);
-	
+
+	/**
+	 * Get boolean `TRUE` if column has primary key or unique key to compute client row id.
+	 * @return bool|null
+	 */
+	public function GetIdColumn ();
+
+	/**
+	 * Set boolean `TRUE` if column has primary key or unique key to compute client row id.
+	 * @param  bool|null $idColumn
+	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
+	 */
+	public function SetIdColumn ($idColumn);
+
 	/**
 	 * Get property automatic formating view helper name.
-	 * @return string|NULL
+	 * @return string|null
 	 */
 	public function GetViewHelper ();
 	
 	/**
 	 * Set property automatic formating view helper name.
-	 * @param  string|NULL $viewHelper
+	 * @param  string|null $viewHelper
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetViewHelper ($viewHelper);
@@ -209,7 +222,7 @@ interface IColumn {
 	 * as integer for pixel value, float for flex value
 	 * or string including `px` or `%` units.
 	 * Width is used only for table grid type.
-	 * @return string|int|float|NULL
+	 * @return string|int|float|null
 	 */
 	public function GetWidth ();
 
@@ -218,7 +231,7 @@ interface IColumn {
 	 * as integer for pixel value, float for flex value
 	 * or string including pixels or percentage value.
 	 * Width is used only for table grid type.
-	 * @param  string|int|float|NULL $width
+	 * @param  string|int|float|null $width
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetWidth ($width);
@@ -227,7 +240,7 @@ interface IColumn {
 	 * Get column min. width, it can be defined as integer 
 	 * or float for pixel value or string including `px` 
 	 * or `%` units. Min. width is used only for table grid type.
-	 * @return string|int|float|NULL
+	 * @return string|int|float|null
 	 */
 	public function GetMinWidth ();
 
@@ -235,7 +248,7 @@ interface IColumn {
 	 * Set column min. width, it can be defined as integer 
 	 * or float for pixel value or string including `px` 
 	 * or `%` units. Min. width is used only for table grid type.
-	 * @param  string|int|float|NULL $minWidth
+	 * @param  string|int|float|null $minWidth
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetMinWidth ($minWidth);
@@ -244,7 +257,7 @@ interface IColumn {
 	 * Get column max. width, it can be defined as integer 
 	 * or float for pixel value or string including `px` 
 	 * or `%` units. Min. width is used only for table grid type.
-	 * @return string|int|float|NULL
+	 * @return string|int|float|null
 	 */
 	public function GetMaxWidth ();
 
@@ -252,7 +265,7 @@ interface IColumn {
 	 * Set column max. width, it can be defined as integer 
 	 * or float for pixel value or string including `px` 
 	 * or `%` units. Min. width is used only for table grid type.
-	 * @param  string|int|float|NULL $maxWidth
+	 * @param  string|int|float|null $maxWidth
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetMaxWidth ($maxWidth);
@@ -262,7 +275,7 @@ interface IColumn {
 	 * integer or float value for `flex-grow` only or as string 
 	 * value for full CSS `flex` shorthand to define `flex-grow`, 
 	 * `flex-shrink` and `flex-basis`.
-	 * @return string|int|float|NULL
+	 * @return string|int|float|null
 	 */
 	public function GetFlex ();
 
@@ -271,14 +284,14 @@ interface IColumn {
 	 * integer or float value for `flex-grow` only or as string 
 	 * value for full CSS `flex` shorthand to define `flex-grow`, 
 	 * `flex-shrink` and `flex-basis`.
-	 * @param  string|int|float|NULL $flex
+	 * @param  string|int|float|null $flex
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetFlex ($flex);
 
 	/**
 	 * Get column additional css classes for head cell and body cell.
-	 * @return \string[]|NULL
+	 * @return \string[]|null
 	 */
 	public function GetCssClasses ();
 
@@ -290,42 +303,42 @@ interface IColumn {
 	public function SetCssClasses ($cssClasses);
 	
 	/**
-	 * Get `TRUE` for editable column, `FALSE|NULL` for not editable column (by default).
-	 * @return bool|NULL
+	 * Get `TRUE` for editable column, `FALSE|null` for not editable column (by default).
+	 * @return bool|null
 	 */
 	public function GetEditable ();
 
 	/**
-	 * Set `TRUE` for editable column, `FALSE|NULL` for not editable column (by default).
-	 * @param  bool|NULL $editable
+	 * Set `TRUE` for editable column, `FALSE|null` for not editable column (by default).
+	 * @param  bool|null $editable
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetEditable ($editable);
 	
 	/**
-	 * Get `TRUE` for disabled column, `FALSE|NULL` for enabled column (enabled by default).
-	 * @return bool|NULL
+	 * Get `TRUE` for disabled column, `FALSE|null` for enabled column (enabled by default).
+	 * @return bool|null
 	 */
 	public function GetDisabled ();
 
 	/**
-	 * Set `TRUE` for disabled column, `FALSE|NULL` for enabled column (enabled by default).
-	 * @param  bool|NULL $disabled
+	 * Set `TRUE` for disabled column, `FALSE|null` for enabled column (enabled by default).
+	 * @param  bool|null $disabled
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetDisabled ($disabled);
 
 	/**
 	 * Get `TRUE` to always send column in AJAX response, no matter if column is active 
-	 * or not, `FALSE|NULL` to not send inactive column (`NULL` by default).
-	 * @return bool|NULL
+	 * or not, `FALSE|null` to not send inactive column (`null` by default).
+	 * @return bool|null
 	 */
 	public function GetAlwaysSend ();
 
 	/**
 	 * Set `TRUE` to always send column in AJAX response, no matter if column is active 
-	 * or not, `FALSE|NULL` to not send inactive column (`NULL` by default).
-	 * @param  bool|NULL $alwaysSend
+	 * or not, `FALSE|null` to not send inactive column (`null` by default).
+	 * @param  bool|null $alwaysSend
 	 * @return \MvcCore\Ext\Controllers\DataGrids\Configs\Column
 	 */
 	public function SetAlwaysSend ($alwaysSend);